fix: Reuse commitment grouping validation.
