+++ src/tools/financial-commitment-reports/create-financial-commitment-report.ts
@@ -4,6 +4,7 @@ import MCPUserError from "../structure/MCPUserError";
 import registerTool from "../structure/registerTool";
 import { pastDateIntervalOptions } from "../utils/dateIntervalOptions";
 import dateValidator from "../utils/dateValidator";
+import { groupingDescription, groupingSchema } from "./schemas";
 
 const description = `
 Create a new Financial Commitment Report in Vantage.
@@ -28,38 +29,6 @@ the full list of available financial_commitments fields and examples.
 `.trim();
 
 type CreateFinancialCommitmentReportRequest = RequestBodyForPathAndMethod<"/v2/financial_commitment_reports", "POST">;
-
-const financialCommitmentGroupings = [
-  "provider",
-  "service",
-  "resource_account_id",
-  "provider_account_id",
-  "commitment_type",
-  "commitment_id",
-  "cost_type",
-  "cost_category",
-  "cost_sub_category",
-  "instance_type",
-  "region",
-] as const;
-
-const groupingDescription =
-  "Grouping dimensions for aggregating financial commitments on the report. Valid groupings: provider, service, resource_account_id, provider_account_id, commitment_type, commitment_id, cost_type, cost_category, cost_sub_category, instance_type, region, and tag:<tag_key>.";
-
-const groupingSchema = z
-  .string()
-  .min(1)
-  .refine(
-    (value) =>
-      financialCommitmentGroupings.includes(value as (typeof financialCommitmentGroupings)[number]) ||
-      value.startsWith("tag:"),
-    {
-      error: groupingDescription,
-      when(payload) {
-        return z.string().min(1).safeParse(payload.value).success;
-      },
-    }
-  );
 
 export default registerTool({
   name: "create-financial-commitment-report",
